Refactor : ChannelModal들에서 중복되는 부분들을 MainModal로 이동
