+++ frontend/src/components/Main/Cam/CamListHeader.tsx
@@ -78,7 +78,11 @@ function CamListHeader(props: CamListHeaderProps): JSX.Element {
           <DropdownMenu
             name="Cam 생성"
             setIsDropdownActivated={setIsDropdownActivated}
-            modalContents={<CreateCamModal />}
+            modalContents={{
+              contents: <CreateCamModal />,
+              title: 'Cam 생성',
+              description: '생성할 Cam의 이름을 작성해주세요',
+            }}
           />
         </Dropdown>
       </CamListHeaderButton>
