+++ frontend/src/components/Main/MainModal.tsx
@@ -2,6 +2,9 @@ import React, { useContext } from 'react';
 import styled from 'styled-components';
 
 import { MainStoreContext } from './MainStore';
+import { BoostCamMainIcons } from '../../utils/SvgIcons';
+
+const { Close } = BoostCamMainIcons;
 
 const Container = styled.div`
   position: fixed;
@@ -24,13 +27,98 @@ const ModalBackground = styled.div`
   background-color: rgb(0, 0, 0, 0.5);
 `;
 
+const Modal = styled.div`
+  width: 35%;
+  min-width: 400px;
+  height: 70%;
+  min-height: 450px;
+
+  background-color: #222322;
+
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  align-items: center;
+
+  border-radius: 20px;
+
+  z-index: 3;
+`;
+
+const ModalInnerBox = styled.div`
+  width: 100%;
+  height: 100%;
+  padding: 30px 10px;
+  display: flex;
+  flex-direction: column;
+  justify-content: flex-start;
+  align-items: flex-start;
+`;
+
+const ModalHeader = styled.div`
+  width: 100%;
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  align-items: center;
+
+  flex: 1;
+`;
+
+const ModalTitle = styled.span`
+  margin-left: 25px;
+  padding: 10px 5px;
+
+  color: #cbc4b9;
+  font-size: 32px;
+  font-weight: 600;
+`;
+
+const ModalDescription = styled.span`
+  flex: 0.3;
+  margin-left: 25px;
+  padding: 10px 5px;
+
+  color: #cbc4b9;
+  font-size: 15px;
+`;
+
+const ModalCloseButton = styled.div`
+  width: 32px;
+  height: 32px;
+  display: flex;
+  flex-direction: center;
+  align-items: center;
+
+  cursor: pointer;
+  margin-right: 25px;
+`;
+
+const CloseIcon = styled(Close)`
+  width: 20px;
+  height: 20px;
+  fill: #a69c96;
+`;
+
 function MainModal(): JSX.Element {
   const { setIsModalOpen, modalContents } = useContext(MainStoreContext);
+  const { contents, title, description } = modalContents;
 
   return (
     <Container>
       <ModalBackground onClick={() => setIsModalOpen(false)} />
-      {modalContents}
+      <Modal>
+        <ModalInnerBox>
+          <ModalHeader>
+            <ModalTitle>{title}</ModalTitle>
+            <ModalCloseButton onClick={() => setIsModalOpen(false)}>
+              <CloseIcon />
+            </ModalCloseButton>
+          </ModalHeader>
+          <ModalDescription>{description}</ModalDescription>
+          {contents}
+        </ModalInnerBox>
+      </Modal>
     </Container>
   );
 }
