+++ frontend/src/types/modal.ts
@@ -1,0 +1,7 @@
+type ModalContents = {
+  contents: JSX.Element;
+  title: string;
+  description: string;
+};
+
+export default ModalContents;
